refactor: reorganize helper functions for better modularity
- Move hasToolCalls() to message-utils.ts for reusability
- Move saveChatTitle() to chat actions where it belongs
- Create message-preparation.ts for message prep logic
- Keep streaming-specific functions in create-chat-stream-response.ts
- Improve code organization and separation of concerns
